requirements.txt
audiosegment==0.22.2
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
decorator==4.4.2
fuzzywuzzy==0.18.0
fuzzywuzzy-stubs==0.0.1
gTTS==2.1.1
gTTS-token==1.1.3
html2text==2020.1.16
htmldate==0.6.3
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
jusText==2.2.0
langdetect==1.0.8
lxml==4.5.1
moviepy==1.0.3
numpy==1.18.4
Pillow==7.1.2
proglog==0.1.9
pydub==0.24.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytube3==9.6.4
readability-lxml==0.7.1
requests==2.23.0
requests-file==1.5.1
six==1.15.0
slackbot==0.5.6
slacker==0.14.0
soupsieve==2.0.1
tldextract==2.2.2
tqdm==4.46.0
trafilatura==0.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
webrtcvad==2.0.10
websocket-client==0.44.0
youtube-transcript-api==0.3.1
